requirements.txt
# Web框架和API
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
requests==2.31.0
aiohttp==3.9.1
flask==3.0.0

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1

# 音频处理
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
pyaudio==0.2.11
numpy==1.24.3
scipy==1.11.4

# 图像处理和OCR
opencv-python==4.8.1.78
pillow==10.1.0
pytesseract==0.3.10

# AI和机器学习
openai==1.3.7
transformers==4.36.2
torch==2.1.1
torchaudio==2.1.1

# TTS语音合成
azure-cognitiveservices-speech==1.34.0
edge-tts==6.1.9
gTTS==2.4.0

# 自动化和浏览器控制
selenium==4.15.2
undetected-chromedriver==3.5.4
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# 数据处理
pandas==2.1.4
pyyaml==6.0.1
python-dotenv==1.0.0

# Web相关
python-multipart==0.0.6
python-socketio==5.10.0
eventlet==0.33.3
gevent==23.9.1

# 系统工具
psutil==5.9.6
schedule==1.2.0
watchdog==3.0.0

# 日志和输出
colorlog==6.8.0
rich==13.7.0

# CLI工具
typer==0.9.0

# Web界面
streamlit==1.28.2
gradio==4.8.0

# 中文处理
jieba==0.42.1
pypinyin==0.50.0

# 音频元数据
mutagen==1.47.0
eyed3==0.9.7
tinytag==1.10.1
python-magic==0.4.27

# 屏幕截图和控制
mss==9.0.1
keyboard==0.13.5
mouse==0.7.1

# 音频播放
playsound==1.3.0
simpleaudio==1.0.4
pygame==2.5.2
python-vlc==3.0.18122

# 开发工具
pytest==7.4.3
black==23.11.0
flake8==6.1.0
